+++ src/main/java/de/hysky/skyblocker/mixins/MinecraftClientMixin.java
@@ -6,9 +6,9 @@ import de.hysky.skyblocker.config.SkyblockerConfigManager;
 import de.hysky.skyblocker.skyblock.item.HotbarSlotLock;
 import de.hysky.skyblocker.skyblock.item.ItemProtection;
 import de.hysky.skyblocker.skyblock.item.SkyblockInventoryScreen;
-import de.hysky.skyblocker.utils.JoinWorldPlaceholderScreen;
-import de.hysky.skyblocker.utils.ReconfiguringPlaceholderScreen;
 import de.hysky.skyblocker.utils.Utils;
+import de.hysky.skyblocker.utils.render.gui.JoinWorldPlaceholderScreen;
+import de.hysky.skyblocker.utils.render.gui.ReconfiguringPlaceholderScreen;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.DownloadingTerrainScreen;
 import net.minecraft.client.gui.screen.ReconfiguringScreen;
