+++ src/main/java/de/hysky/skyblocker/skyblock/chat/filters/AdFilter.java
@@ -1,4 +1,4 @@
-package de.hysky.skyblocker.skyblock.filters;
+package de.hysky.skyblocker.skyblock.chat.filters;
 
 import de.hysky.skyblocker.config.SkyblockerConfigManager;
 import de.hysky.skyblocker.utils.Constants;
