+++ src/main/java/de/hysky/skyblocker/skyblock/chat/filters/SimpleChatFilter.java
@@ -1,4 +1,4 @@
-package de.hysky.skyblocker.skyblock.filters;
+package de.hysky.skyblocker.skyblock.chat.filters;
 
 import de.hysky.skyblocker.utils.chat.ChatPatternListener;
 import net.minecraft.text.Text;
