+++ src/test/java/de/hysky/skyblocker/skyblock/chat/filters/AbilityFilterTest.java
@@ -1,4 +1,4 @@
-package de.hysky.skyblocker.skyblock.filters;
+package de.hysky.skyblocker.skyblock.chat.filters;
 
 import org.junit.jupiter.api.Test;
